Add tables for cpu/gpu + handle different provider versions
