+++ shared/dbSchemas/akash/index.ts
@@ -3,6 +3,8 @@ export { DeploymentGroup } from "./deploymentGroup";
 export { Provider } from "./provider";
 export { ProviderSnapshot } from "./providerSnapshot";
 export { ProviderSnapshotNode } from "./providerSnapshotNode";
+export { ProviderSnapshotNodeCPU } from "./providerSnapshotNodeCPU";
+export { ProviderSnapshotNodeGPU } from "./providerSnapshotNodeGPU";
 export { ProviderAttribute } from "./providerAttribute";
 export { ProviderAttributeSignature } from "./providerAttributeSignature";
 export { Lease } from "./lease";
